Suppression des relations
